Log in/requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
pytest==8.0.2
python-dotenv==1.0.0 